+++ torchy_baselines/common/evaluation.py
@@ -47,8 +47,8 @@ def evaluate_policy(model, env, n_eval_episodes=10, deterministic=True,
     mean_reward = np.mean(episode_rewards)
     std_reward = np.std(episode_rewards)
     if reward_threshold is not None:
-        assert mean_reward > reward_threshold, 'Mean reward below threshold: '\
-                                         '{:.2f} < {:.2f}'.format(mean_reward, reward_threshold)
+        assert mean_reward > reward_threshold, (f'Mean reward below threshold: '
+                                                '{mean_reward:.2f} < {reward_threshold:.2f}')
     if return_episode_rewards:
         return episode_rewards, n_steps
     return mean_reward, std_reward